requirements.txt
flask==2.3.3
flask-login==0.6.2
flask-wtf==1.1.1
werkzeug==2.3.7
firebase-admin==6.2.0
cloudinary==1.33.0
python-dotenv==1.0.0
Pillow==10.0.0
email-validator==2.0.0
pyrebase4==4.7.1
flask-migrate==4.0.4
wtforms==3.0.1
requests==2.31.0
gunicorn==21.2.0
flask-cors
